Fix 64-to-32 integer truncating. This is a bit more of a problem with byte-swapping than just reading/writing.

+++ netcon/mtclient/mtgametrack.cpp
@@ -332,7 +332,7 @@ void RequestGameList() {
 
   int len = strlen(DLLPXO_hosted_lobby_name) + 1;
   memcpy(GameListReq.data, DLLPXO_hosted_lobby_name, len);
-  GameListReq.len = INTEL_INT(GAME_HEADER_ONLY_SIZE + len);
+  GameListReq.len = INTEL_INT((uint32_t)(GAME_HEADER_ONLY_SIZE + len));
   SendGameTrackerPacker(&GameListReq);
 }
